+++ event.py
@@ -3,13 +3,12 @@
 import json
 from myfunc import log
 
-class Event:
-    name = ''
-    time = ''
-    game = ''
-    
+class Event:    
     def __init__ (self, name):
         self.name = name
+        self.time = ''
+        self.game = ''
+        self.members = {}
         
     def updateStream(self, time):
         self.time = time
@@ -19,10 +18,15 @@ class Event:
         self.time = time
         self.game = game
         self.save()
+
+    def addMember(self, member):
+        self.members[member.id] = member.display_name
+        self.save()
         
     def reset(self):
         self.time = ''
         self.game = ''
+        self.members = {}
         self.save()
             
     def save(self):
@@ -32,11 +36,14 @@ class Event:
     def load(self):
         try:
             with open(self.name + '.json', 'r') as f:
-                self.time = json.load(f)['time']
+                data = f.read()
+                self.time = json.loads(data)['time']
                 log(f'Event-Zeit für {self.name} wurde geladen: {self.time}')
-                self.game = json.load(f)['game']
-                log(f'Event-Name für {self.name} wurde geladen: {self.game}')
-
+                if self.name == 'game':
+                    self.game = json.loads(data)['game']
+                    log(f'Event-Game für {self.name} wurde geladen: {self.game}')
+                self.members = json.loads(data)['members']
+                log(f'Event-Members für {self.name} wurden geladen: {self.members}')
         except:
             pass
             log(f'Event {self.name} konnte nicht geladen werden und bleibt bei: {self.time} - {self.game}')